Fix blocking on redis connection.
Each check on a failed node was blocking the connection indefinitely, impacting on the others nodes monitored.
The redis client used didnt allowed to change the timeout, so i started using Jedis for it.

I have not introduced a new test case for it, but it is easy to reproduce, adding a rule on iptables like:
 `sudo iptables -A INPUT -p tcp --destination-port 6379 -j DROP`